doc: indent use-package sexp + remove ensure t
* Avoiding ensure and straight in the same use-package sexp

The use of the `:ensure` and `:straight` keywords in the same use-package call
is discouraged by the author of straight (see
https://github.com/raxod502/straight.el/issues/425#issuecomment-546706730) and
thus I propose removing the =:ensure t= bit to avoid confusion.

* Formatting with a focus on readability

Furthermore, I just aligned the sexp within the block (which removed some
leading whitespace 🤷🏿‍♂️ (sorry for the larger diff). What is perhaps more
useful, is the space that I added before `:repo` to trick the emacs indenter to
format the sexp in a manner that more clearly outlines the relationship between
the keywords (siblings, same level) which kind of gets lost in the older
indentation.

```elisp
(use-package org-hyperscheduler
  :ensure t
  :straight
  (:repo "dmitrym0/org-hyperscheduler"
         :host github
         :type git
         :files ("*")))
```

I considered rephrasing the value of the `straight` keyword to the _long-form_:

```elisp
(org-hyperscheduler :repo "dmitrym0/org-hyperscheduler"
                    :host github
                    :type git
                    :files ("*"))
```

instead of the current _short-form_:

```elisp
( :repo "dmitrym0/org-hyperscheduler"
  :host github
  :type git
  :files ("*"))
```

but since the straight README
https://github.com/raxod502/straight.el#integration-with-use-package-1 outlines
that the _short-form_ macroexpands to contain the _package name_ (which looks
more like the _long-form_, I have no reason to be difficult and poke the bear
so, I'm leaving this alone.