Add autoload magic comments for `org-hyperscheduler-open` functions
These magic commands are expanded to autoload-function-calls when the package is
compiled. They allow the main interactive entry functions to be before the
package has been loaded. I.e. this means that one can add `:defer t` to the
use-package declaration and then have it be loaded lazily only once one of the
autoloaded commands is called. Alternatively, the user could also add them to
the `:commands` use-package keyword, however I prefer doing that for them for
the main commands.

Also I made org-hs-open interactive, which I think it's meant to be as it's not
used anywhere else in the code and thus probably just an alias? Anyway, this
will probably be better solved by my other PR #6, possibly we might have to
solve a merge-conflict when merging both, but I don't think it's a problem.

Also for main functions that I touched in this PR I added a dot to the end of
the docstring because my emacs checkdoc checker (which I use via flycheck)
complained the docstring should end with a dot.